tracer: add members for current and last instruction to state::State
Currently, `tracer` code queries instructions from the `Binary`
repeatedly, which is a failible operation that potentially has some cost
associated to it. Since we are interested in just the current and last
instruction (at least most of the time), we may as well keep them in the
state.

This change introduces just the data members and initialization. We will
start making use of them in future changes.