tracer: add state::State fn for computing inferable jump target
We plan to encapsulate anything concerning instruction tracing for a
single packet to `state::State`. This includes the computation of
inferable jump targets. This is currently done directly in
`Tracer::next_pc`, but we intend to replace this fn with a new
`state::State` fn in the future. Encapsulating this computation in a fn
allows for more expressive code.