tracer: move fn for determining the next PC to state::State
We plan to encapsulate anything concerning instruction tracing for a
single packet to `state::State`. This obviously includes determining the
next PC.

We take this opportunity to rewrite `next_pc`, making the logic more
obvious and removing the now unnecessary `Payload` parameter. We also
make it return the next PC and `Instruction` in an `Item` alongside the
`stop_here` flag the old implementation returned.