tracer: add member for current stack depth
The `tracer` module contains the tracing algorithm, which is mostly a
translation of pseudo-code from the Efficient Trace for RISC-V document.
The result is not particularly ideomatic Rust. Part of that is loops
reporting events via callbacks.

We plan to switch form the existing collection of reporting callbacks to
an `Iterator` as the primary reporting interface in the future. We thus
need to make sure all data that currently is just kept on in a local
variable from one loop iteration to the next is preserved properly. This
includes the stack depth which is currently retreived from the current
`Payload`.