types: add type encapsulating information on traps
We intend to extend information on traps that is reported. This change
introduces a struct for bundling that information so we can pass it
through more conveniently. However, we chose to omit the EPC from this
struct. It will not be availible in all situations we intend to use
`trap::Info` in.