tracer: add type for transporting trace data
The `tracer` module contains the tracing algorithm, which is mostly a
translation of pseudo-code from the Efficient Trace for RISC-V document.
The result is not particularly ideomatic Rust. Part of that is loops
reporting events via callbacks.

We plan to switch form the existing collection of reporting callbacks to
an `Iterator` as the primary reporting interface in the future. This
change introduces a type we can use as `Iterator::Item`. It should catch
anything we deem worthwile to report. This differs from what is
currently reported by the existing callbacks.