types: add a module for RISC-V specific types
The `decoder` module contains quite a few types that are not exactly
specific to tracing packets and payloads. Some of them are specific to
RISC-V in general while others are associated to generic concepts and
shared between `decoder` and `tracer`.

We don't want those pinned down in `decoder`, especially not types that
are used as data memebers of data types in `tracer`. We thus decided to
move those to a separate module. This change introduces that module.