decoder: add module for decoder utilities
We have a few utilities for decoding specific items, most of them in the
form of a fn. They currently live in `payload` module or close to the
type (that doesn't implement `Decode` itself) they handle.

We want to free some types from the need to reside in `decoder`, and we
also want to reduce the size of the `payload` module. Moving some
utilities to a dedicated module helps achieve both of those goals.