decoder: add fn for creating a branch::Map
We want to free some types from the need to reside in `decoder`. This
includes `branch::Map`, which is currently constructed exclusively by
`branch::Count`. Either of those types is to be moved, we cannot
construct them the same way any more. So, in preparation, we add a fn
that is public within the crate for creating `branch::Map`s.