tracer: add state::State member for stop condition
The `tracer` module contains the tracing algorithm, which is mostly a
translation of pseudo-code from the Efficient Trace for RISC-V document.
The result is not particularly ideomatic Rust. Part of that is loops
reporting events via callbacks.

We plan to switch form the existing collection of reporting callbacks to
an `Iterator` as the primary reporting interface in the future. For
this, we need to replace current mechanisms through which tracing loops
are broken. Currently, loops are broken based on a number of conditions
that depend on the current `Payload`. This change introduces another
vehicle for those conditions. Since it concerns when to stop tracing
instructions, we place it in `state::State`.