tracer: move Error to dedicated module
The `tracer` module houses the tracing algorithm, which is mostly a
translation of pseudo-code from the Efficient Trace for RISC-V document.
The result is not particularyl ideomatic Rust with quite a few too-long
fns. We therefore wish to modularize it more.

This change prepares for that by moving `Error` into a dedicated `error`
submodule. This makes is a bit less awkward to refer to this type from
other submodules in the future. It also makes `tracer.rs` a bit smaller.