tracer: initialize stop_at_last_branch when receiving a full branch map
We are supposed to stop following the execution path at either a given
address (which is included in the packet) or the last branch recorded.
The latter is the case if we received a full branch map and the address
is not set in the `Payload::Branch`.

This was correct initially, but we errorneously removed the logic to set
this particular flag when we migrated to branch maps in

    e6d1274f7651fbb7056d52d91f1db3b949ecd5ef