tracer: use state::StopCondition in Tracer::follow_execution_path
The `tracer` module contains the tracing algorithm, which is mostly a
translation of pseudo-code from the Efficient Trace for RISC-V document.
The result is not particularly ideomatic Rust. Part of that is loops
reporting events via callbacks.

We plan to switch form the existing collection of reporting callbacks to
an `Iterator` as the primary reporting interface in the future. For
this, we need to replace current mechanisms through which tracing loops
are broken. We recently introduces a `state::State` member to this end.
This change now migrates `Tracer::follow_execution_path` from a set of
free conditions based the current `Payload` to a `state::StopCondition`
initialized from that `Payload`.