tracer: add state::State fn for querying whether the state is fused
The `tracer` module contains the tracing algorithm, which is mostly a
translation of pseudo-code from the Efficient Trace for RISC-V document.
The result is not particularly ideomatic Rust. Part of that is loops
reporting events via callbacks.

We plan to switch form the existing collection of reporting callbacks to
an `Iterator` as the primary reporting interface in the future. For this
purpose, we recently introduced representations of loop breaking
conditions, including a fused state in which we shall not advance the
`state::State`. This change instroduces a fn which makes determining
this condition more convenient.