tracer: add Tracer data member for iteration state
The `tracer` module contains the tracing algorithm, which is mostly a
translation of pseudo-code from the Efficient Trace for RISC-V document.
The result is not particularly ideomatic Rust. Part of that is loops
reporting events via callbacks.

We plan to switch form the existing collection of reporting callbacks to
an `Iterator` as the primary reporting interface in the future. For this
purpose, we recently introduced representations of loop breaking
conditions. However, we currently have two (albeit similar) loops and
one additional location in which items are reported. In a future impl of
`Iterator::next`, we need to know which of those states we are in.

This change introduces a data member for recording this information,
which also happens to include information that was previously kept in
`Tracer::start_of_trace`.