tracer: initialize priviledge before reporting Items
The `tracer` module contains the tracing algorithm, which is mostly a
translation of pseudo-code from the Efficient Trace for RISC-V document.
The result is not particularly ideomatic Rust. Part of that is loops
reporting events via callbacks.

We plan to switch form the existing collection of reporting callbacks to
an `Iterator` as the primary reporting interface in the future. Users
will pass a packet to `Tracer::process_te_inst` and then iterate over
`Item`s for that packet.

This change prepares for the transition by making sure the reporting
loops are the last thing happening in a `Tracer::process_te_inst` call.
We can more this initialization because (a) we don't report the
privilege and (b) the existing checks during the iteration do not clash
since they are only active for version 1 of the tracing protocol.