tracer: eliminate Tracer::get_instr
The `tracer` module contains the tracing algorithm, which, in the past,
was mostly a translation of pseudo-code from the Efficient Trace for
RISC-V document. The result was not particularly ideomatic Rust. Part of
that is loops reporting events via callbacks. We recently introduced an
`Iterator` interface and cut back reporting via those callbacks and
currently plan to remove them alltogether.

`Tracer::get_instr` did nothing but call `Binary::get_insn` and report
the retrieved instructionon success. Given that the only a single call
site remains, we decided to eliminate it and call `Binary::get_insn`
directly.