tracer: remove callback based reporting from Tracer
The `tracer` module contains the tracing algorithm, which, in the past,
was mostly a translation of pseudo-code from the Efficient Trace for
RISC-V document. The result was not particularly ideomatic Rust. Part of
that is loops reporting events via callbacks. We recently introduced an
`Iterator` interface and cut back reporting via those callbacks.

This change removes the remaining reporting calls as well as the
`Tracer::report_trace` data member from `Tracer`.