tracer: remove ReportTrace
The `tracer` module contains the tracing algorithm, which, in the past,
was mostly a translation of pseudo-code from the Efficient Trace for
RISC-V document. The result was not particularly ideomatic Rust. Part of
that is loops reporting events via callbacks. We recently introduced an
`Iterator` interface and removed any reporting calls from `Tracer`.
Thus, `ReportTrace` no longer serves a purpose.