tracer: add error::Error variant for unprocessed instructions
The `tracer` module contains the tracing algorithm, which is mostly a
translation of pseudo-code from the Efficient Trace for RISC-V document.
The result is not particularly ideomatic Rust. Part of that is loops
reporting events via callbacks.

We plan to switch form the existing collection of reporting callbacks to
an `Iterator` as the primary reporting interface in the future. Users
will pass a packet to `Tracer::process_te_inst` and then iterate over
`Item`s for that packet.

We'll rely on users to exhaust all availible tracing `Item`s via the
iteration before feeding the net packet. This change introduces an
`error::Error` variant for when they don't.