instruction: add fn and impls for extracting Bits from data
`Bits` used to have a (private) fn for extracting values from a
`Segment`. We removed it in

    e1580358f9fad39e0bdbf64fbadb0b6316e66fc4

That was around the time we introduced `binary::Binary` and ripped out
the old method for supplying `Instruction`s. The fn was, broadly
speaking, in the way of remodelling (removing `Segment`).

Still, `Instruction`s will need to be extracted from byte slices at some
point by some users. And we don't want to burden them too much with
figuring this out themselves.