instruction: add fn and impls for extracting an Instruction from data
We recently introduced a fn for extracting `Bits` from a byte slice. It
is intended for future implementations of `Binary`. `Binary::get_insn`,
however, needs to return an `Instruction`, not `Bits`. The conversion is
trivial, but still needs to be done, and we don't want to burden
implementors with it.