tracer: add initializer for state::State
For historical reasons, `Tracer` does have access to `state::State`'s
internals and makes regular use of that for initializing a state for
instruction tracing. For reasons that should be obvious, we want
`state::State` to be way more opaque. We want `state::State` to only
expose a few fns for driving the instruction tracing, some query fns and
an interface that allows initialization of the state as long as it is
not fused. We consider initialization of an unfused state an error.

This change instroduces such an interface which enforces this rule.
We will exclusively use this interface for initialization in the future.
This will ultimately allow us to make the `state::State` data members
private.