tracer: use match for sync differentiation in tracer::process_te_inst
Previously, we would differentiate bewteen the various `Synchronization`
variants via an if-else chain. This is a remnant of being a transcript
of pseudo-code from the specification.

Given that we need to cover all variants one way or another, and that
there are more than two of them, we prefer a `match` over an if-else
chain.