tracer: set address whenever its present in Tracer::process_te_inst
Originally, our representation of branch maps was much closer to the
pseudo code from the specification: a lot of the burden of handling
packages was on the `Tracer`, including the differentiation between a
regular and "full" branch map and the associated presence and
non-presence of an address in the packet.

However, we since introduced an abstraction for branch maps and put
addresses in `Option`s where is may not be present. It appears we messed
up the logic for retrieving potentially relative addresses during the
migration. Among other things, this made us run into a panic in
`decoder::payload::Payload::get_address` for full branch maps.