tracer: use state::Initializer in Tracer::sync_init
For historical reasons, `Tracer` does have access to `state::State`'s
internals and makes regular use of that for initializing a state for
instruction tracing. For reasons that should be obvious, we want
`state::State` to be way more opaque. For this purpose, we recently
introduced `state::Initializer`, a safe interface for initializing a
`state::State`.

This change adapts `Tracer::sync_init` and repurposes it by making it
return a `state::Initializer`.