tracer: use state::Initializer in Tracer::process_te_inst
For historical reasons, `Tracer` does have access to `state::State`'s
internals and made regular use of that for initializing a state for
instruction tracing. For reasons that should be obvious, we want
`state::State` to be way more opaque. For this purpose, we recently
introduced `state::Initializer`, a safe interface for initializing a
`state::State`.

This change adapts `Tracer::process_te_inst`, also removing the need for
the initial explicit check for a fused `Tracer::state`, since that is
done by the `State::initializer` (whereever it ends up being called).