tracer: do not query State::pc directly in Tracer
For historical reasons, `Tracer` does have access to `state::State`'s
internals and made regular use of that for initializing a state for
instruction tracing. However, we recently migrated to a new interface
for that initialization, removing any need for that access.

The only site left where `Tracer` does still access a `state::State`
data member was for determining the EPC. This change makes `Tracer` use
a query fn instead.