tracer: make state::State data members private
For historical reasons, `Tracer` does have access to `state::State`'s
internals and made regular use of that for initializing a state for
instruction tracing. For reasons that should be obvious, we want
`state::State` to be way more opaque. For this purpose we removed any
access to `state::State` data members from `Tracer`. This change now
reaps the rewards of those efforts by making the data members private.