tracer: integrate setting stop condition in existing control flow
We recently moved the remaining stop condition logic of the now gone
`Tracer::follow_execution_path` into `Tracer::process_te_inst`. This
lead to some repetition in the control flow.

This change merges the bits together, removing the need for a mutable
flag. The logic is not _strictly_ the same as before, but the cases
where it differed (no branch or address packet) only concern extensions
that the tracer does not (yet) implement anyway.