lib: introductory rewrite example for Iterator
Originally, we `tracer::Tracer` would drive instruction tracing and
report via callbacks. However, we transitioned to an `Iterator` based
interface and ripped out the old reporting callbacks. While we kept the
example in compilable, it lost most of any expressiveness it once had.

This change rewrites the example using a real-world (!) machine code
snippet: a simple copy loop. We chose to not have any comments in the
example but add some flavor text instead.