tracer: use state::Initializer in Tracer::process_support
For historical reasons, `Tracer` does have access to `state::State`'s
internals and made regular use of that for initializing a state for
instruction tracing. For reasons that should be obvious, we want
`state::State` to be way more opaque. For this purpose, we recently
introduced `state::Initializer`, a safe interface for initializing a
`state::State`.

This change adapts `Tracer::process_support`.