tracer: make state::State fns take a reference to a mutable Binary
We decided to make `Binary::get_insn` take a mutable reference to `self`
rather than an immutable one, as this allows for greater freedom in
impls' design. In preparation for this, we need to ensure that we do
have a mutable `Binary` (or reference) everywhere we need to query an
`Instruction`.