instruction: make binary::Binary::get_insn take &mut self
When introducing the `Binary` trait, we decided to make it not require
any mutable operations. After all, we are just querying `Instruction`s.
Still, if we want to implement any form of caching (as we did in the
past in `Tracer`), we need to alter some internal state (the cache).

We can have inner mutable state, but we are restricted by our decision
to support `nostd` contexts. Besides, a given `Tracer` will query one
instruction after the other (parallelism of any sort is not feasible)
and never leak its `Binary`.

We therefore deem it acceptable to require a mutable `Binary` for
operations. On the other hand, making `Binary::get_insn` take a mutable
reference to self allows for greater freedom in `Binary` design.