instruction: add Binary impl for ELF files
Most users will probably trace programs they have an ELF file of. And we
ourselfes will likely also benefit from being able to handle ELF files
for tests. The ability to use the contents of an ELF as a `Binary` is of
value to us.

This change introduces a `Binary` for static, uncompressed ELF files. We
considered just implementing `Binary` for `ElfBytes`, but decided that
even though we cannot perform allocation for abstracting data structures
(in the `nostd` context) we still at least don't want to walk through
all segments every time we look for an instruction. Instead, we cache
the last segment we ended up using and check whether or not it is
relevant before any other segment.