examples: add a simple tracer program
We already have a small example in the crate's top level documetnation.
However, an example developers can execute directly, with any input they
chose, is nicer in our case. In addition, we can use such an example for
some (manual) end-to-end testing.

Thus, we introduce an example that we can throw against trace files and
programs in the reference flow's test suites.