deps: add elf as an optional dependency
Most users will probably trace programs they have an ELF file of. And we
ourselfes will likely end up generating traces for testing by running an
ELF file through a simulator. The ability to use the contents of an ELF
as a `Binary` is of value to us.

For obvious reasons, we don't want to parse ELF files ourselves but use
an existing parser. We decided on the `elf` crate, which is also usable
in `nostd` contexts.