deps: add serde as an optional dependency
Users will often want to read configuration from some file, especially
the protocol configuration. We can make this easier for them by making
some configuration types impl `serde::Deserialize`. For that, we need to
have it as a dependency. And since some users may not want to have
`serde` as a dependency on their end, we choose to make it an optional
dependency.