deps: add toml as a dev-dependency
We plan to introduce an example tracer both to show off the crate's
capabilities and for testing agains the reference flow. For that, we
might need to read encoder parameters from a file. The reference flow
uses what the python `configparser` things is an INI file. TOML is close
enough to perform the necessary conversion with `egrep`.