instruction: add Binary moved by a fixed offset
Sometimes, a `Binary` doesn't map absolute virtual addresses to
`Instruction`s. Instead, a `Binary` may map code to a section starting
at `0`. This includes PIE ELF files. In such cases, we need to remap the
entire `Binary`.

We could solve this issue for each and every affected `Binary` impl, but
the issue is common enough that we decided to have a simple wrapper.