tracer: fuse state::State when hitting end with no inferred address
In the absence of an inferred address, we are supposed to halt tracing
for the current packet if we hit a code-dependent end-condition (which
is cimmunicated by the return value of `state::State::next_pc`).

The orignal code, which was transcribed from pseudo-code provided in the
specification, this was done via a fn return. When we transitioned to an
`Iterator` interface, we needed to translate that into fusing the state.
It appears we forgot to do so.