config: move default values for Protocol to constant
In the past, we did have a constant defining default `Protocol` values.
However, we then implemented `Default` instead. But while `Default` is
more useful in practice, it is also much more opaque when it comes to
documentation. We thus decided to re-introduce the constant (and use it
for implementing `Default::default)