config: reorder Protocol data members to reflect order in spec
We want to make `Protocol` a proper representation of _encoder_
parameters as defined in Chapter 10. Parameters and Discovery of the
specification. That chapter contains tables of parameters to the
encoder. We'll probably make looking for specific parameters easier for
everyone if the order in our struct is the same as in those tables.