types: make trap::Info::tval an Option expressing the kind of trap
We only need to differentiate between interrupts and exceptions, since
that's the only trap types existing for RISC-V. We also only need to
carry `tval` for the latter. In fact, we only _should_ carry it in the
latter case since the value will be nonsensical for interrupts anyway.

We therefore decided to make `trap::Info::tval` an `Option` that will be
`None` for interrupts and `Some` value for exceptions. And this makes
`trap::Info::kind` redundant, which is why we remove it in favor of
query fns we already have.