decoder: move Synchronization Payload represenation to new submodule
The `payload` submodule grew quite sizable, partially due to the
relatively huge number of types it holds. Since the synchronization
payloads are somewhat removed from the rest of the packets, and also
make up a significant portion of those types, we decided to move them to
a separate module. We took the opportunity to extend the documentation
of the types we moved.