decoder: add data member for hart index widths
Currently, we supply the hart index width to the `Decoder` via
`config::Protocol::cpu_index_width`. However, the hart index is part of
the packet `header::Header`. And in the future, we want to constrain
`config::Protocol` to data relevant for decoding `Payload`s.

We therefore decided to move the hart index width away from
`config::Protocol` to `Decoder` itself.