decoder: use Decoder::hart_index_width for header::Header
Previously, we supplied the hart index width to the `Decoder` via
`config::Protocol::cpu_index_width`. However, the hart index is part of
the packet `header::Header`. And in the future, we want to constrain
`config::Protocol` to data relevant for decoding `Payload`s.

We therefore recently introduced a new data member to `Decoder. This
change now makes use of that member for decoding `header::Header`.