decoder: use enum for sync::Support::encoder_mode rather than u64
`sync::Support::encoder_mode` identifies a specific mode the encoder is
operating in. Rather than a bare integer, we prefer an enum for this
kind of information. And given that, according to the current version
(2.0.3) of the spec, there is exactly one valid value, the `u64`
previously used was especially wasteful.

Since the width of the field in encoded packets is not fixed by the
specification, we choose to construct the enum value via `TryFrom<u8>`
rather than via `Decode`.