decoder: make Decoder carry a Unit
Support packets carry fields which the specification says are (trace
unit) implementation specific regarding their widths and contents.
Previously, we coped with that using configurable widths, and by just
ignoring the ioptions field.

In order to decode support packages with meaningful `ioptions`, we
recently introduced the `unit::Unit` trait. This change now makes the
`Decoder` carry a `unit` which we will use for decoding support
packages in the future. We choose to not add a trait bound to `Decoder`
itself or its main impl block in order to keep `Decode` impls simple.
After all, the bound will only be relevant for `sync::Support`.