decoder: decode sync::Support::encoder_mode and ioptions using Unit
Support packets carry fields which the specification says are (trace
unit) implementation specific regarding their widths and contents.
Previously, we coped with that using configurable widths, and by just
ignoring the ioptions field.

In order to decode support packages with meaningful `ioptions`, we
recently introduced the `unit::Unit` trait. This change now makes use of
that trait for decoding `sync::Support`. Which, as a result, gains a
type parameter for the `ioptions`. And so does `sync::Synchronization`
and `payload::Payload`.

We choose to use `unit::ReferenceIOptions` as a default. Thus, we do not
need to adapt the `tracer` module. However, this also means that
`tracer::Tracer` fns currently only accept "reference" packets.