decoder: make Builder::build consume Builder
We plan to introduce new `Decoder` data members in the future which will
most likely be prepared within the `Builder`. Copying or cloning those
may not be viable, or rather, we don't want to make such an assumption.
Given that `tracer::Builder::build` does consume `self` and it is more
unlikely that we instantiate more than one `Decoder`, we deemed it well
acceptable for `Builder::build` to consume `self`. `Builder` itself will
still impl `Copy` and `Clone` as long as all its members do.