config: remove config::Protocol members for support field widths
Support packets carry fields which the specification says are (trace
unit) implementation specific regarding their widths and contents.
Previously, we coped with that using configurable widths, and by just
ignoring the ioptions field.

In order to decode support packages with meaningful `ioptions`, we
recently introduced the `unit::Unit` trait. This trait features a fn for
querying the `encoder_mode` width and a fn for decoding `ioptions`.
Thus, we do not need to provide those widths via the config any more.