config: move AddressMode from tracer to config
The address mode is a piece of configuration that is of most interest to
the `tracer` module. However, it may also communicated by the trace unit
and thus show up in `decoder::unit::Unit::IOptions` in some form. If we
want to represent them properly, we better move them out of the `tracer`
module.