decoder: add trait for ioptions
We recently introduced `unit::Unit` for expressing specifics about trace
unit implementations. That includes the associated type `IOptions` which
expresses the options active for the trace (from that point on).

Currently, we do not honour those options. In fact, the options are not
even accessible in a generic, implementation-agnostic way. This change
introduces a trait as a solution for the latter.