decoder: impl unit::IOptions for unit::ReferenceIOptions
We recently introduced `unit::Unit` for expressing specifics about trace
unit implementations. That includes the associated type `IOptions` which
expresses the options active for the trace (from that point on).

We recently introduced a new trait for as an abstraction of types we
want to use as `unit::Unit::IOptions`. This change now implements this
trait for the `unit::ReferenceIOptions`.