decoder: mandate unit::Options for unit::Unit::IOptions
We recently introduced `unit::Unit` for expressing specifics about trace
unit implementations. That includes the associated type `IOptions` which
expresses the options active for the trace (from that point on).

We recently introduced a new trait for as an abstraction of types we
want to use as `Unit::IOptions`. This change mandates that trait be
implemented for all those types.