tracer: use variables for Payloads in tests
Currently, `Tracer` fns accepting any kind of `Payload` only accepts
ones from the reference trace unit. We plan to make them accept any
`Payload` with `decoder::unit::IOptions`. This change prepares for that
by making sure the type of the payloads is known and does not need to be
inferred from the fn call itself.