config: add fn for creating an AddressMode from a flag
We may often run into situations in which the address mode is provided
as a flag that must be converted to the enum. This is trivial, but still
requires a few lines if formatted properly.

This could also be done in a `From<bool>` impl, but we deemed that to be
too ambiguous with regards to the mapping from boolean values to address
modes.