tracer: add error::Error variant for unsupported features
A trace unit may communicate active options via a support packet's
ioptions field. Previously, we would not honour those options, but we
plan to change that. We may come across options that we do not (yet)
implement. In such cases, we cannot simply continue tracing but need to
signal an error.