tracer: control address computation with explicit address mode
In the past, address computation was already controlled by a `Tracer`
data member of type `AddressMode`. However, back then we still lacked
the sign-extension logic. When we introduced that in

    08ff9c1c2bd8ef23ede10964f33a74abc599059c

we replaced the `AddressMode` with an `Option` containing the bits in
the address field we needed to extend in the delta address mode. We did
that because we thought that the configuration was static and fixed from
the `Tracer`'s creation. However, options may be set via support
packets, and thus the `Tracer` needs the ability to alter the address
_after_ its initial creation.

We thus bring back the address mode member. We do keep the logic and
data member for the sign-extension, since we need it for delta address
mode. However, we now keep the width directly, not behind an `Option`.