tracer: add state::Initializer fn for controlling sequential jumps
We recently introduced `unit::Unit` for expressing specifics about trace
unit implementations. That includes the associated type `IOptions` which
expresses the options active for the trace (from that point on).

Currently, we do not honour those options, which may specify whether
sequentially inferable jumps are reported by the trace unit or we need
to infer them. We already have the `state::State::sequential_jumps` data
member, but no possibility to set it after `state::State` creation.