tracer: add state::State member controlling implicit returns
Previously, we considered the implicit return option to be active
whenever we had a return stack that could actually hold any return
addresses. However, despite what the parameters allow for, the implicit
return feature may be controlled via the ioptions field of support
packages.

This change therefore introduces a new flag for controlling whether to
infer function returns and a `state::Initializer` fn for setting that
flag.