decoder: add struct encapsulating various field widths
Currently, `Decoder` carries a `config::Protocol` from which we derive
field widths in a number of `Decode` impls and utility fns. In the
future, we want to make that type a proper representation of _encoder_
parameters. This will undoubtly increase it's size, making it a bit more
awkward to carry it around in a `Decoder`.

We thus decided to make `Decoder` carry a different type `Widths`, which
only contains the information we actually need for decoding payloads.
This will also involve moving some minor calculations from utility fns
into a `From` impl we use to create `Widths` from `config::Protocol`.

We also took the opportunity to reflect optionality of fields via
`core::num::NonZeroU8` and `Option`. This made it necessary to unwrap
a few `Option`s for now, but those failible operations will be removed
when we adapt `config::Protocol` in the future.