decoder: make Builder::with_config take a reference
Currently, `Decoder` carries a `config::Protocol` from which we derive
field widths in a number of `Decode` impls and utility fns. In the
future, we want to make that type a proper representation of _encoder_
parameters. This will undoubtly increase it's size, making it a bit more
awkward to copy.

We plan to make `Decoder` and `Builder` not carry a `config::Protocol`
but a more specialized type. We don't want to copy a potentially large
config type just to deconstruct it in oder to construct a new type.
Instead, we want to operate on a reference and extract whatever we need.
We prepare for these future changes by adapting the (public) `Builder`
interface first.