decoder: make Decoder carry Widths rather than config::Protocol
Currently, `Decoder` carries a `config::Protocol` from which we derive
field widths in a number of `Decode` impls and utility fns. In the
future, we want to make that type a proper representation of _encoder_
parameters. This will undoubtly increase it's size, making it a bit more
awkward to carry it around in a `Decoder`.

We thus decided to make `Decoder` carry a different type `Widths`, which
only contains the information we actually need for decoding payloads.