config: make Protocol parameters NonZeroU8 where 0 is invalid
We want to make `Protocol` a proper representation of _encoder_
parameters as defined in Chapter 10. Parameters and Discovery of the
specification. For many of those parameters, a value of `0` is invalid.

Previously, we used that invalid value to signal absence for
`context_width_p` and `time_width_p`. In order to preserve the ability
to specify the absence of context or time information, we add the
paremters `nocontext_p` and `notime_p` described by the specification.