tracer: make Builder carry individual values rather than config
Previously, `Builder` carried a `config::Protocol` from which we derived
various settings for the `Tracer`. However, in the future, we want to
make `config::Protocol` a proper representation of _encoder_ parameters.
That will undoubtly increase it's size, making it a bit more awkward to
carry around, especially considering that we only need a few pieces of
information. We therefore decided make `Builder` not carry a
`config::Protocol` but already derived settings.