tracer: make Builder::with_config take a reference
Currently, `Builder` carries a `config::Protocol` from which we derive
various settings for the `Tracer`. In the future, we want to make
`config::Protocol` a proper representation of _encoder_ parameters. This
will undoubtly increase it's size, making it a bit more awkward to carry
around.

We therefore decided make `Builder` not carry a `config::Protocol` but
already derived settings. We don't want to copy a potentially large
config type just to deconstruct it. Instead, we want to operate on a
reference and extract the few pieces of information we actually need for
configuring the `Tracer`.