config: (de)serialize bool params in Protocol as 0 or 1
We want to make `Protocol` a proper representation of _encoder_
parameters as defined in Chapter 10. Parameters and Discovery of the
specification. Accoring to the specification, those parameters are
always expressed as numbers, even those that are effectively `bool`.

For obvious reasons, we want to keep our `bool`s, but for the sake of
compatability, we want them to (de)serialize from/to `0` or `1`. For
this purpose, this change introduces a `config` submodule holding some
serde utilities. Deserialization is known to work with `toml`. Other
formats/`Deserializer` impls may require some tweaks.