examples: make simple tracer print debug ouput if DEBUG env var is 1
We introduced the example tracer partially because for comparing our
trace reconstruction to the reference flow's decoder model. That
features a debug flag for printing more information. We decided to
introduce something similar so we can compare our putput with theirs
(after some filtering).

But since we don't want to introduce a proper argument parser, we use an
environment variable instead...