examples: enable simple tracer to compare against a reference
We can already compare the simple tracer's default ouptut against both
the output of the reference flow's decoder model and the ground truth.
However, that's only the PCs. We can also compare our debug output
against that of the decoder model, but that's also not ideal.

At least in the reference flow, the ground truth is a spike trace in the
form of a CSV. We can easily compare against that and abort on any
mismatch. We only miss out on the EPCs.

Since we don't want to pull in another dependency just for processing
CSVs, we get our hands dirty doing that by hand.