decoder: move Packet to smi submodule
A `Packet` with an `smi::Header` clearly is an `smi::Packet`. We thus
make it so. Besides, being more specific now helps mitigate future name
clashes.

On top of that, we also cleaned up some documentation that needed some
changes anyway due to the move and the subsequent removal of the
`Payload` import from the toplevel `decoder` module.