decoder: test decoding extensions with correct f0s width
Currently, the subformat field of format 0 payloads is always decoded
from exactly one bit. However, the specification states that the width
of that subformat is controlled by the parameter `f0s_width_p` with a
default value of `0`.

We plan to make that parameter control the width as specified. We also
intend to use the default stated in the specification. Therefore, these
tests will fail with the default parameters in the future.