decoder: add Error variant for unknown formats
Currently, the subformat field of format 0 payloads is always decoded
from exactly one bit. However, the specification states that the width
of that subformat is controlled by the parameter `f0s_width_p`.

We plan to make that parameter control the width as specified. Thus, it
becomes possible for raw packets to carry subformats that are unknown to
us. This change introduces an `Error` variant to communicate such
failures.